requirements.txt
Django==4.2.11
django-braces==1.16.0
django-model-utils==5.0.0
django-star-ratings==0.9.2